% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_utilities.R
\name{all.refs.valid}
\alias{all.refs.valid}
\title{Checks that all SBtab references are valid}
\usage{
\method{all}{refs.valid}(tab, allvars = all.vars(tab, reac = FALSE))
}
\arguments{
\item{tab}{a list of tables}

\item{allvars}{a list of all variable IDs}
}
\description{
A reference is a column name that starts with ">": >Calcium is a
reference to a compound (probably) called Calcium (or a Parameter,
etc.)
}
