% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_utilities.R
\name{all.vars}
\alias{all.vars}
\title{collect all variable names}
\usage{
\method{all}{vars}(tab, react = TRUE)
}
\arguments{
\item{tab}{a list of lists with the table content}

\item{react}{if TRUE, this function only collects the IDs of
Compouds, Parameters and Expressions}
}
\value{
a vector of all names
}
\description{
this function collects all varibale names in SBtab, here we use the
!ID column (because the !Name column can be something that isn't a
programming language kind of variable-name).
}
\details{
If the react flag is FALSE then really all names are collected,
otherwise only those that can appear in reaction formulae and
kinetic laws of reactions.
}
