% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_to_sbml.R
\name{.interpret.unit.from.string}
\alias{.interpret.unit.from.string}
\title{Unit Interpreter}
\usage{
.interpret.unit.from.string(unit.str)
}
\arguments{
\item{unit.str}{a string that contains a human readable unit}
}
\value{
data.frame with an interpretation of the unit
}
\description{
This function will try its best to interpret strings like
"liter/(nmol ms)"
rules: 1. only one slash is allowed
       2. M can be mega or mol/l
       3. prefixes and units can be words or single letters
       4. everything after a slash is the denominator
       5. u is an accepted replacement for μ
       6. no parentheses
this retruns a data.frame with components as in the sbml standard:
kind, multiplier, scale and exponent
since there is only one slash,parentheses do nothing
everything after a slash is the denominator, so: l/mol s is the same as (l)/(mol s)
Remark: not all units are understood.
}
