% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_to_sbml.R
\name{.make.sbml}
\alias{.make.sbml}
\title{Create an SBML model from Tables}
\usage{
.make.sbml(
  H,
  Defaults,
  Constant = NULL,
  Parameter,
  Input = NULL,
  Expression = NULL,
  Reaction,
  Compound,
  Output,
  Comp
)
}
\arguments{
\item{H}{model header, a Model identifying string}

\item{Defaults}{A data.frame with default units}

\item{Constant}{Table of Constants for this model [optional]}

\item{Parameter}{Table of Parameters}

\item{Input}{Table of input parameters, treated the same as normal Parameters}

\item{Expression}{will be interpreted and transformed into somthing that SBML understands (optional)}

\item{Reaction}{Table of reactions}

\item{Compound}{Table of species (reacting compounds)}

\item{Output}{Table of output functions (a model of measurement/observables)}

\item{Comp}{Table of Compartments}
}
\value{
SBML model as a character vector
}
\description{
This function creates a character vector that contains the
biological model in SBML form. The return value can be written into
a file to create an SBML document.
}
\details{
The input is an SBtab model as data.frames
}
