% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listop.R
\name{\%otherwise\%}
\alias{\%otherwise\%}
\title{otherwise operator}
\usage{
a \%otherwise\% b
}
\arguments{
\item{a}{anything}

\item{b}{a default value, also anything}
}
\value{
a; but if is.null(a), return b
}
\description{
Returns the first operand, unless it evaluates to NULL, then the
second is returned.
}
