% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexop.R
\name{\%s\%}
\alias{\%s\%}
\title{regular expression substitution}
\usage{
text \%s\% pat
}
\arguments{
\item{text}{text to match against pattern}

\item{pat}{regular expression, and replacement}
}
\value{
a vector of the sam elength as text, with substituted items
}
\description{
This simplifies the syntax of regular expression matching.
}
\examples{
> text <- c('a','b','c')
> text \%s\% c('[ab]','_')
[1] "_" "_" "c"
}
