% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexop.R
\name{\%~\%}
\alias{\%~\%}
\title{regular sub-expression matching}
\usage{
text \%~\% pattern
}
\arguments{
\item{text}{text to match against pattern}

\item{pattern}{regular expression pattern (can have subexpressions)}
}
\value{
a list of the same length as text, with lists of matches in each item
}
\description{
This simplifies the syntax of regular expression matching.
}
\examples{
match some math expressions:
x <- c('2*x^2','x','3')
x \%~\% '([0-9]*)[*]?x?(\\\\^[0-9]+)?'
[[1]]
[1] "2*x^2" "2"     "^2"

[[2]]
[1] "x" ""  ""

[[3]]
[1] "3" "3" ""
}
