% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_utilities.R
\name{id.eq.name}
\alias{id.eq.name}
\title{Are the IDs the same as the Names}
\usage{
id.eq.name(T)
}
\arguments{
\item{T}{one table (data.frame)}
}
\description{
This is not mandatory, but there could be errors if somewhere in
the document a Name is used inside of mathematical expressions
while the ID was needed. This can help find those mistakes.
}
