% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_utilities.R
\name{make.cnames}
\alias{make.cnames}
\title{Make C compatible names}
\usage{
make.cnames(Labels)
}
\arguments{
\item{Labels}{a character vector with words that need to be turned
into names.}
}
\value{
a vector with unique names, can be used as variable names
 in C
}
\description{
Uses make.names internally, but replaces dots with underscores.
}
