% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_to_mod.R
\name{NeuronUnit}
\alias{NeuronUnit}
\title{A unit string compatible with NEURON}
\usage{
NeuronUnit(unit)
}
\arguments{
\item{unit}{a string}
}
\value{
string with MOD unit (NEURON)
}
\description{
NEURON has additional rules for units, this converts a unit-string
to comply with the rules.
}
