% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_to_mod.R
\name{OneOrMoreLines}
\alias{OneOrMoreLines}
\title{Writes MOD variable definitions}
\usage{
OneOrMoreLines(Prefix, Table, Suffix)
}
\arguments{
\item{Prefix}{a prefix to print for every line}

\item{Table}{a data.frame with variable definitions; we use only the row.names}

\item{Suffix}{a suffix to print for every line}
}
\value{
a character vector with strings in MOD format
}
\description{
If the number of definitions is low, they all appear in one line
Otherwise, this prints one definition per line
}
