% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_to_vfgen.R
\name{paste_tag}
\alias{paste_tag}
\title{paste_tag makes an XML tag from a data-frame}
\usage{
paste_tag(Name, Attributes, indent = " ")
}
\arguments{
\item{Name}{the nameof the printed tag (" <Name [...]>")}

\item{Attributes}{the data.frame containing the tag attributes to print (as strings)}

\item{indent}{the printed string will be prefixed with this}
}
\value{
a character vector, one tag per row of Attributes argument
}
\description{
given a data.frame, this function returns one string per row, using
the data.frame column names as attribute names of an XML tag.
}
\examples{
x <- data.frame(Name="a",Type="b",Formula="c")
tag <- paste_tag("Example",x)
message(tag) #  <Example Name="a" Type="b" Formula="c"/>
}
