% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_utilities.R
\name{sbtab.data}
\alias{sbtab.data}
\title{Read data from SBtab}
\usage{
sbtab.data(tab)
}
\arguments{
\item{tab}{a list of data.frames with SBtab content}
}
\description{
This function reads all datasets from SBtab Documents that are
referenced in the Experiment table and returns them as a list of
data.frames
}
\details{
The returned list always describes simulation experiments:
time series experiments, possibly with scheduled events.
}
