% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_utilities.R
\name{sbtab.header.value}
\alias{sbtab.header.value}
\title{Read a property from the SBtab header}
\usage{
sbtab.header.value(sbtab.header, key = "Document")
}
\arguments{
\item{sbtab.header}{the first line of an SBtab file}

\item{key}{the left side of each key=value pair}
}
\value{
value of the key=value pair
}
\description{
This function retrieves a property from an sbtab header:
PropertyName='PropertyValue' (properties are these key=value
pairs).
}
