% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_utilities.R
\name{sbtab.valid}
\alias{sbtab.valid}
\title{validate spelling in SBtab document}
\usage{
sbtab.valid(tab)
}
\arguments{
\item{tab}{a list of lists as returned by `sbtab_from_tsv()`}
}
\description{
This function checks the SBtab document for consistency.
It finds misspelled varibale names in the reaction kinetics.
}
