% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_to_vfgen.R
\name{sbtab_from_ods}
\alias{sbtab_from_ods}
\title{A parser for .ods files with SBtab document structure}
\usage{
sbtab_from_ods(ods.file)
}
\arguments{
\item{ods.file}{a string (file's name)}
}
\value{
SBtab
a list of tables (data.frames), one per ods sheet
SBtab[['TableName']] retrives a data.frame
comment(SBtab) is the name of the document
}
\description{
This uses the readODS package to read the file.
SBtab sheets are themselves tables dedicated to a specific type of
model property: Reaction, Compound, Parameter, etc.
}
\details{
The SBtab content is not interpreted in any way.
}
\examples{
model.sbtab<-sbtab_from_ods('model.ods')
}
\keyword{import}
