% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_to_vfgen.R
\name{sbtab_from_tsv}
\alias{sbtab_from_tsv}
\title{A parser for a bunch of .tsv files with SBtab document content}
\usage{
sbtab_from_tsv(tsv.file)
}
\arguments{
\item{tsv.file}{a character vector (file names, one per sheet)}
}
\value{
L
a list with two elements:
L[['Document']] is the document name
L[['Table']] is a list with one entry per SBtab sheet
}
\description{
This function reads the files (not using any package).
SBtab sheets are themselves tables dedicated to a specific type of
model property: Reaction, Compound, Parameter, etc.
}
\details{
The SBtab content is not interpreted in any way.
}
\examples{
model.sbtab<-sbtab_from_tsv(dir(pattern='.*[.]tsv$'))
}
\keyword{import}
