% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_utilities.R
\name{sbtab_quantity}
\alias{sbtab_quantity}
\title{read vector from sbtab Quantity table}
\usage{
sbtab_quantity(Table)
}
\arguments{
\item{Table}{the sbtab table imported via sbtab_from_{ods,tsv}
(either)}
}
\value{
a vector with names corresponding to !ID and values taken
    from !DefaultValue, !Value, or !InitialValue
}
\description{
This function extracts a vector from an sbtab table, and names the
elements according to the row names of the table. The value column can be named any of these:
Value, DefaultValue, InitialValue, Mean, Median.
}
\keyword{quantity}
\keyword{sbtab}
