% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_to_vfgen.R
\name{sbtab_to_vfgen}
\alias{sbtab_to_vfgen}
\title{SBtab content exporter}
\usage{
sbtab_to_vfgen(SBtab, cla = TRUE)
}
\arguments{
\item{SBtab}{a list as returned from any of the import functions
[sbtab_from_tsv()] or [sbtab_from_ods()]}

\item{cla}{a logical value that determines whether conservation law
analysis should be performed. If this is TRUE (default) the
resulting ODE model will be reduced to a state variable sub-set
that is independent in the sense of linear algebraic
relationships.}
}
\value{
the text (as a character array) that was also written to
    the .vf file
}
\description{
This function interprets the SBtab content and exports it to three
possible formats: (1) vfgen's `vf` format (xml); (2) NEURON's mod
file; (3) SBML if the libSBML package can be loaded.
}
