% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_utilities.R
\name{update_from_table}
\alias{update_from_table}
\title{copy vector n times and change the columns according to (n×m) Table}
\usage{
update_from_table(v, Table)
}
\arguments{
\item{v}{a vector with named elements}

\item{Table}{an sbtab table with columns that reference members of
v via the ">" prefix (e.g. the column ">ATP" will change the
v['ATP'] value)}
}
\value{
M, a matrix with n columns, where n is the number of rows
    in Table.
}
\description{
Update a vector v to a matrix M with several copies of that vector
with some values changed accoring to a table (Qantity Matrix) If
the table has columns that name members of v (also named) they will
be used. The M will have as many columns as the table has rows.
}
\keyword{matrix}
\keyword{quantity}
\keyword{sbtab}
