% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexop.R
\name{printf}
\alias{printf}
\title{printf prints}
\usage{
printf(...)
}
\arguments{
\item{...}{arguments to sprintf}
}
\value{
nothing
}
\description{
a wrapper around sprintf but it actually outputs on screen.
}
