% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtab_to_vfgen.R
\name{sbtab_from_ods}
\alias{sbtab_from_ods}
\title{A parser for .ods files with SBtab document structure}
\usage{
sbtab_from_ods(ods.file)
}
\arguments{
\item{ods.file}{a string (file's name)}
}
\value{
L
a list with two elements:
L[['Document']] is the document name
L[['Table']] is a list with one entry per SBtab sheet
}
\description{
This uses the readODS package to read the file.
SBtab sheets are themselves tables dedicated to a specific type of
model property: Reaction, Compound, Parameter, etc.
}
\details{
The SBtab content is not interpreted in any way.
}
\examples{
model.sbtab<-sbtab_from_ods('model.ods')
}
\keyword{import}
